move the pagewrapper into components
